# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ContentEditionResponse": {
        "properties": {
          "business_date": {
            "format": "date",
            "title": "Business Date",
            "type": "string"
          },
          "scoring_profile": {
            "title": "Scoring Profile",
            "type": "string"
          },
          "slot_mode_enabled": {
            "title": "Slot Mode Enabled",
            "type": "boolean"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/ContentEditionSlotResponse"
            },
            "title": "Slots",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "business_date",
          "timezone",
          "scoring_profile",
          "slot_mode_enabled",
          "slots"
        ],
        "title": "ContentEditionResponse",
        "type": "object"
      },
...
      "ContentSlotRunResponse": {
        "properties": {
          "acquisition_run_id": {
            "format": "uuid",
            "title": "Acquisition Run Id",
            "type": "string"
          },
          "business_date": {
            "format": "date",
            "title": "Business Date",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "config_fingerprint": {
            "title": "Config Fingerprint",
            "type": "string"
          },
          "content_slot": {
            "enum": [
              "morning",
              "noon",
              "evening"
            ],
            "title": "Content Slot",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "eligible_scores": {
            "title": "Eligible Scores",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "governance_run_id": {
            "format": "uuid",
            "title": "Governance Run Id",
            "type": "string"
          },
          "governed_event_cutoff": {
            "format": "date-time",
            "title": "Governed Event Cutoff",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_limit": {
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Item Limit",
            "type": "integer"
          },
          "preparation_at": {
            "format": "date-time",
            "title": "Preparation At",
            "type": "string"
          },
          "scores_url": {
            "title": "Scores Url",
            "type": "string"
          },
          "scoring_profile": {
            "title": "Scoring Profile",
            "type": "string"
          },
          "selected_count": {
            "title": "Selected Count",
            "type": "integer"
          },
          "slot_policy_fingerprint": {
            "title": "Slot Policy Fingerprint",
            "type": "string"
          },
          "slot_policy_version": {
            "title": "Slot Policy Version",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "target_at": {
            "format": "date-time",
            "title": "Target At",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "total_scores": {
            "title": "Total Scores",
            "type": "integer"
          },
          "trigger": {
            "enum": [
              "manual",
              "scheduled"
            ],
            "title": "Trigger",
            "type": "string"
          },
          "unfilled_count": {
            "title": "Unfilled Count",
            "type": "integer"
          },
          "unfilled_reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Unfilled Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "trigger",
          "business_date",
          "timezone",
          "content_slot",
          "display_name",
          "scoring_profile",
          "acquisition_run_id",
          "governance_run_id",
          "governed_event_cutoff",
          "config_fingerprint",
          "slot_policy_version",
          "slot_policy_fingerprint",
          "preparation_at",
          "target_at",
          "expires_at",
          "item_limit",
          "status",
          "total_scores",
          "eligible_scores",
          "selected_count",
          "unfilled_count",
          "unfilled_reason_codes",
          "error_code",
          "created_at",
          "started_at",
          "completed_at",
          "status_url",
          "scores_url"
        ],
        "title": "ContentSlotRunResponse",
        "type": "object"
      },
      "ContentSlotScoreListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentSlotScoreResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "ContentSlotScoreListResponse",
        "type": "object"
      },
...
      "CopyGenerationDetailResponse": {
        "properties": {
          "active_draft_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Draft Version Id"
          },
          "business_date": {
            "format": "date",
            "title": "Business Date",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content_slot_selection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Slot Selection Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "daily_topic_selection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Topic Selection Id"
          },
          "decision_kind": {
            "enum": [
              "selected",
              "no_topic"
            ],
            "title": "Decision Kind",
            "type": "string"
          },
          "detail_url": {
            "title": "Detail Url",
            "type": "string"
          },
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/CopyDraftResponse"
            },
            "title": "Drafts",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "no_topic_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "No Topic Code"
          },
          "origin_kind": {
            "enum": [
              "legacy_daily",
              "content_slot"
            ],
            "title": "Origin Kind",
            "type": "string"
          },
          "repair_count": {
            "title": "Repair Count",
            "type": "integer"
          },
          "scoring_profile": {
            "title": "Scoring Profile",
            "type": "string"
          },
          "selected_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Event Id"
          },
          "selected_event_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Event Version Id"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "no_topic",
              "accepted",
              "review_required",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin_kind",
          "daily_topic_selection_id",
          "content_slot_selection_id",
          "business_date",
          "timezone",
          "scoring_profile",
          "decision_kind",
          "selected_event_id",
          "selected_event_version_id",
          "no_topic_code",
          "status",
          "active_draft_version_id",
          "repair_count",
          "error_code",
          "created_at",
          "started_at",
          "completed_at",
          "status_url",
          "detail_url",
          "drafts"
        ],
        "title": "CopyGenerationDetailResponse",
        "type": "object"
      },
...
      "CreateContentSlotRunRequest": {
        "properties": {
          "business_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Date"
          },
          "content_slot": {
            "enum": [
              "morning",
              "noon",
              "evening"
            ],
            "title": "Content Slot",
            "type": "string"
          }
        },
        "required": [
          "content_slot"
        ],
        "title": "CreateContentSlotRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for authoritative-source acquisition, factual governance, and auditable event organization, deterministic daily topic selection, and private brand knowledge ingestion with internal copy-generation context retrieval. It does not expose public brand search, publishing, or arbitrary URL-fetch capabilities.",
    "title": "Edu AI Lead Agent API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/content-editions/{business_date}": {
      "get": {
        "operationId": "read_content_edition_api_v1_content_editions__business_date__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Business Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "default": "preview",
              "maxLength": 40,
              "minLength": 1,
              "title": "Profile",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentEditionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Content Edition",
        "tags": [
          "content-slots"
        ]
      }
    },
    "/api/v1/content-slot-runs": {
      "post": {
        "operationId": "create_content_slot_run_api_v1_content_slot_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentSlotRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentSlotRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Content Slot Run",
        "tags": [
          "content-slots"
        ]
      }
    },
    "/api/v1/content-slot-runs/{run_id}": {
      "get": {
        "operationId": "read_content_slot_run_api_v1_content_slot_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentSlotRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Content Slot Run",
        "tags": [
          "content-slots"
        ]
      }
    },
    "/api/v1/content-slot-runs/{run_id}/scores": {
      "get": {
        "operationId": "read_content_slot_scores_api_v1_content_slot_runs__run_id__scores_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentSlotScoreListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Content Slot Scores",
        "tags": [
          "content-slots"
        ]
      }
    },
...
    "/api/v1/copy-generation-runs/{run_id}/detail": {
      "get": {
        "operationId": "read_copy_generation_detail_api_v1_copy_generation_runs__run_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopyGenerationDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Copy Generation Detail",
        "tags": [
          "copy-generation"
        ]
      }
    },
...
    "/api/v1/material-packages/{package_id}": {
      "get": {
        "operationId": "read_material_package_api_v1_material_packages__package_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPackageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Material Package",
        "tags": [
          "material-packages"
        ]
      }
    },
...
  }
}